+++ relay-event-normalization/src/normalize/span/exclusive_time.rs
@@ -141,7 +141,7 @@ pub fn compute_span_exclusive_time(event: &mut Event) {
 #[cfg(test)]
 mod tests {
     use chrono::{TimeZone, Utc};
-    use relay_event_schema::protocol::{EventType, Timestamp, TraceId};
+    use relay_event_schema::protocol::{EventType, Timestamp};
     use similar_asserts::assert_eq;
 
     use super::*;
@@ -159,7 +159,7 @@ mod tests {
             contexts: {
                 let mut contexts = Contexts::new();
                 contexts.add(TraceContext {
-                    trace_id: Annotated::new(TraceId("4c79f60c11214eb38604f4ae0781bfb2".into())),
+                    trace_id: Annotated::new("4c79f60c11214eb38604f4ae0781bfb2".parse().unwrap()),
                     span_id: Annotated::new(SpanId(span_id.into())),
                     ..Default::default()
                 });
@@ -183,7 +183,7 @@ mod tests {
             description: Annotated::new(description.into()),
             start_timestamp: Annotated::new(start),
             timestamp: Annotated::new(end),
-            trace_id: Annotated::new(TraceId("4c79f60c11214eb38604f4ae0781bfb2".into())),
+            trace_id: Annotated::new("4c79f60c11214eb38604f4ae0781bfb2".parse().unwrap()),
             span_id: Annotated::new(SpanId(span_id.into())),
             parent_span_id: Annotated::new(SpanId(parent_span_id.into())),
             ..Default::default()