+++ relay-event-normalization/src/trimming.rs
@@ -436,6 +436,7 @@ fn slim_frame_data(frames: &mut Array<Frame>, frame_allowance: usize) {
 mod tests {
     use std::iter::repeat_n;
 
+    use crate::MaxChars;
     use chrono::DateTime;
     use relay_event_schema::protocol::{
         Breadcrumb, Context, Contexts, Event, Exception, ExtraValue, SentryTags, Span, SpanId,
@@ -443,8 +444,6 @@ mod tests {
     };
     use relay_protocol::{get_value, Map, Remark, SerializableAnnotated};
     use similar_asserts::assert_eq;
-
-    use crate::MaxChars;
 
     use super::*;
 
@@ -1035,7 +1034,7 @@ mod tests {
     #[test]
     fn test_untrimmable_fields() {
         let original_description = "a".repeat(819163);
-        let original_trace_id = TraceId("b".repeat(48));
+        let original_trace_id: TraceId = "b".repeat(32).parse().unwrap();
         let mut event = Annotated::new(Event {
             spans: Annotated::new(vec![
                 Span {
@@ -1044,7 +1043,7 @@ mod tests {
                 }
                 .into(),
                 Span {
-                    trace_id: original_trace_id.clone().into(),
+                    trace_id: original_trace_id.into(),
                     ..Default::default()
                 }
                 .into(),
@@ -1067,7 +1066,7 @@ mod tests {
     fn test_untrimmable_fields_drop() {
         let original_description = "a".repeat(819164);
         let original_span_id = SpanId("b".repeat(48));
-        let original_trace_id = TraceId("c".repeat(48));
+        let original_trace_id: TraceId = "c".repeat(32).parse().unwrap();
         let original_segment_id = SpanId("d".repeat(48));
         let original_op = "e".repeat(129);
 
@@ -1080,7 +1079,7 @@ mod tests {
                 .into(),
                 Span {
                     span_id: original_span_id.clone().into(),
-                    trace_id: original_trace_id.clone().into(),
+                    trace_id: original_trace_id.into(),
                     segment_id: original_segment_id.clone().into(),
                     is_segment: false.into(),
                     op: original_op.clone().into(),