+++ relay-server/src/envelope/mod.rs
@@ -618,7 +618,6 @@ fn require_termination(slice: &[u8], offset: usize) -> Result<(), EnvelopeError>
 #[cfg(test)]
 mod tests {
     use relay_base_schema::project::{ProjectId, ProjectKey};
-    use uuid::Uuid;
 
     use super::*;
 
@@ -1124,7 +1123,7 @@ mod tests {
     #[test]
     fn test_parametrize_root_transaction() {
         let dsc = DynamicSamplingContext {
-            trace_id: Uuid::new_v4(),
+            trace_id: "67e5504410b1426f9247bb680e5fe0c8".parse().unwrap(),
             public_key: ProjectKey::parse("abd0f232775f45feab79864e580d160b").unwrap(),
             release: Some("1.1.1".to_string()),
             user: Default::default(),